# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AXYNTRAX Dashboard Supremo — API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Stream a chat completion from Claude or Gemini (SSE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
...
  /finances:
    get:
      operationId: listFinances
      tags: [finance]
      summary: Lista movimientos financieros
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [ingreso, egreso] }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finance"
    post:
      operationId: createFinance
      tags: [finance]
      summary: Crea un movimiento financiero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceCreate"
      responses:
        "201":
          description: Movimiento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finance"
  /finances/{id}:
    delete:
      operationId: deleteFinance
      tags: [finance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Eliminado
...
components:
  schemas:
...
    AiChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
      required: [role, content]
    AiChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiChatMessage"
        provider:
          type: string
          enum: [claude, gemini]
      required: [messages]
...
    Finance:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [ingreso, egreso] }
        amount: { type: number }
        currency: { type: string }
        category: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        clientId: { type: ["integer", "null"] }
        paymentId: { type: ["integer", "null"] }
        date: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, type, amount, currency, date, createdAt]
    FinanceCreate:
      type: object
      properties:
        type: { type: string, enum: [ingreso, egreso] }
        amount: { type: number }
        currency: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        clientId: { type: ["integer", "null"] }
        date: { type: ["string", "null"], format: date-time }
      required: [type, amount]
...